+++ src/faker_file/__init__.py
@@ -1,5 +1,5 @@
 __title__ = "faker_file"
-__version__ = "0.12.3"
+__version__ = "0.12.4"
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
 __copyright__ = "2022-2023 Artur Barseghyan"
 __license__ = "MIT"